# api/swagger/swagger.yaml
...
swagger: '2.0'
info:
  description: ''
  version: '0.0.1'
  title: 'data'
  termsOfService: ''
  contact:
    name: ''
    url: ''
    email: 'steven.taylor@me.com'
  license:
    name: 'Steven Taylor'
host: localhost
basePath: /
schemes:
- https
consumes:
- application/json
produces:
- application/json
securityDefinitions:
  Oauth:
    type: oauth2
    scopes:
      open: no security
      user: a user
    flow: implicit
    authorizationUrl: https://my_aouth/
security:
  - Oauth: []
paths:
...
  /alarm/{id}/mode/arm/{mode}:
    x-swagger-router-controller: alarm
    get:
      operationId: setMode
      security:
        - Oauth:
            - user
      parameters:
        - in: path
          name: id
          required: true
          type: string
        - in: path
          name: mode
          required: true
          type: string
          enum: [
            stay,
            away
          ]
      responses:
        200:
          description: ok
          schema:
            type: object
        default:
          $ref: '#/responses/Error'
  /alarm/{id}/chime/{state}:
    x-swagger-router-controller: alarm
    get:
      operationId: setChimeMode
      security:
        - Oauth:
            - user
      parameters:
        - in: path
          name: id
          required: true
          type: string
        - in: path
          name: state
          required: true
          type: string
          enum: [
            on,
            off
          ]
      responses:
        200:
          description: ok
          schema:
            type: object
        default:
          $ref: '#/responses/Error'
...
responses:
  Error:
    description: error
    schema:
      $ref: '#/definitions/Error'
definitions:
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
...